refactor(config): move public base URL helpers to utils.urls
